Revert "CLASSES AND LOGIN/REGISTER SYSTEM DONE"
