+++ AdminSystem/1Viewer.aspx
@@ -1,4 +1,4 @@
-﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="3Viewer.aspx.cs" Inherits="_1Viewer" %>
+﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="1Viewer.aspx.cs" Inherits="_1Viewer" %>
 
 <!DOCTYPE html>
