--- AdminSystem/Dashboard.aspx
@@ -1,24 +1,0 @@
-﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="Dashboard.aspx.cs" Inherits="_1_List" %>
-
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head runat="server">
-   <title>Dashboard</title>
-    <style>
-        .logout-button {
-            position: absolute;
-            top: 10px;
-            right: 10px;
-        }
-    </style>
-</head>
-<body>
-    <form id="form1" runat="server">
-        <div>
-            <h2>Welcome, <asp:Label ID="lblUserName" runat="server" Text=""></asp:Label></h2>
-            <asp:Button ID="btnLogout" runat="server" Text="Logout" CssClass="logout-button" OnClick="btnLogout_Click" />
-        </div>
-    </form>
-</body>
-</html>
