--- AdminSystem/Login.aspx
@@ -1,22 +1,0 @@
-﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="Login.aspx.cs" Inherits="_1Viewer" %>
-
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head runat="server">
-  <title>Login</title>
-</head>
-<body>
-    <form id="form1" runat="server">
-        <div>
-            <h2>Login</h2>
-            <asp:Label ID="lblEmail" runat="server" Text="Email:"></asp:Label>
-            <asp:TextBox ID="txtEmail" runat="server"></asp:TextBox><br />
-            <asp:Label ID="lblPassword" runat="server" Text="Password:"></asp:Label>
-            <asp:TextBox ID="txtPassword" runat="server" TextMode="Password"></asp:TextBox><br />
-            <asp:Button ID="btnLogin" runat="server" Text="Login" OnClick="btnLogin_Click" /><br />
-            <asp:Label ID="lblMessage" runat="server" Text=""></asp:Label>
-        </div>
-    </form>
-</body>
-</html>
