--- AdminSystem/RegisterUser.aspx
@@ -1,25 +1,0 @@
-﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="RegisterUser.aspx.cs" Inherits="_1_DataEntry" %>
-
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head runat="server">
-    <title>Register User</title>
-</head>
-<body>
-    <form id="form1" runat="server">
-        <div>
-            <h2>Register User</h2>
-            <asp:Label ID="lblUserName" runat="server" Text="User Name:"></asp:Label>
-            <asp:TextBox ID="txtUserName" runat="server"></asp:TextBox><br />
-            <asp:Label ID="lblPassword" runat="server" Text="Password:"></asp:Label>
-            <asp:TextBox ID="txtPassword" runat="server" TextMode="Password"></asp:TextBox><br />
-            <asp:Label ID="lblEmail" runat="server" Text="Email:"></asp:Label>
-            <asp:TextBox ID="txtEmail" runat="server"></asp:TextBox><br />
-          
-            <asp:Button ID="btnRegister" runat="server" Text="Register" OnClick="btnRegister_Click" /><br />
-            <asp:Label ID="lblMessage" runat="server" Text=""></asp:Label>
-        </div>
-    </form>
-    </body>
-    </html>
