+++ ClassLibrary/ClassLibrary.csproj
@@ -43,8 +43,6 @@
   <ItemGroup>
     <Compile Include="clsCar.cs" />
     <Compile Include="clsDataConnection.cs" />
-    <Compile Include="ClsUser.cs" />
-    <Compile Include="ClsUsersCollection.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
